[Refactor] #6: SpaceService를 BuildingService와 병합
